+++ config/prod.exs
@@ -1,3 +1,3 @@
-use Mix.Config
+import Config
 
 config :td_df, TdDfWeb.Endpoint, server: true