+++ test/support/authentication.ex
@@ -16,10 +16,8 @@ defmodule TdDfWeb.Authentication do
     |> put_req_header("authorization", "Bearer #{jwt}")
   end
 
-  def create_user_auth_conn(%{role: role} = claims) do
-    {:ok, jwt, full_claims} = Guardian.encode_and_sign(claims, %{role: role})
-    {:ok, claims} = Guardian.resource_from_claims(full_claims)
-    register_token(jwt)
+  def create_user_auth_conn(%{} = claims) do
+    %{jwt: jwt, claims: claims} = authenticate(claims)
 
     conn =
       ConnTest.build_conn()
@@ -39,31 +37,13 @@ defmodule TdDfWeb.Authentication do
     }
   end
 
-  def build_user_token(%Claims{} = claims) do
-    case Guardian.encode_and_sign(claims, %{}) do
-      {:ok, jwt, _full_claims} -> register_token(jwt)
-      _ -> raise "Problems encoding and signing a user"
-    end
-  end
+  defp authenticate(%{role: role} = claims) do
+    {:ok, jwt, %{"jti" => jti, "exp" => exp} = full_claims} =
+      Guardian.encode_and_sign(claims, %{role: role})
 
-  def build_user_token(user_name, opts \\ []) when is_binary(user_name) do
-    user_name
-    |> create_claims(opts)
-    |> build_user_token()
-  end
-
-  def get_user_token(user_name) do
-    user_name
-    |> build_user_token()
-    |> register_token
-  end
-
-  defp register_token(token) do
-    case Guardian.decode_and_verify(token) do
-      {:ok, _} -> :ok
-      _ -> raise "Problems decoding and verifying token"
-    end
-
-    token
+    {:ok, claims} = Guardian.resource_from_claims(full_claims)
+    {:ok, _} = Guardian.decode_and_verify(jwt)
+    TdCache.SessionCache.put(jti, exp)
+    %{jwt: jwt, claims: claims}
   end
 end