+++ test/test_helper.exs
@@ -1,5 +1,6 @@
 {:ok, _} = Application.ensure_all_started(:ex_machina)
 
+TdCache.Redix.del!()
 ExUnit.start()
 
 Ecto.Adapters.SQL.Sandbox.mode(TdDf.Repo, :manual)