+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-3614] Support for access token revocation
+
 ## [4.47.0] 2022-07-04
 
 ### Added