+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [5.3.0] 2023-03-13
 
 ### Added
 
-- [TD-3806] Hierarchy template cache implementation 
+- [TD-3806] Hierarchy template cache implementation
 
 ## [5.0.0] 2023-01-30